Refactor Makefile to support cross-compilation
This makes several changes, but the result is mostly the same for host
builds. It makes it possible to crosscompile and trivially so under
Nerves by letting Make variables specify where system libraries and
header files are located.

The most notable change is that the shared library is under the priv
directory rather than a subdirectory with the architecture. When
crosscompiling, the target architecture (at least how Erlang refers to
it) is not nearly as easy to determine.

Finally, the Elixir NIF loading code was updated with the new path to
the shared library and a autoload turned off to avoid errors when
compiling.